requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-analytics-data==0.17.0
litellm==1.17.0
pandas==2.1.3
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
gspread==5.12.0